requirements.txt
Django==3.2.16
djangorestframework==3.13.1
